search box completed

+++ src/Action/LoginAction.js
@@ -1,28 +1,42 @@
 import axios from 'axios'
 import BrowserHistory from '../Components/Utils/BrowserHistory';
 
-export function registerHandle(payload){
-const options = {
-url: 'http://localhost:4090/Signin',
-method: 'POST',
-data: payload
-};
+export function registerHandle(payload) {
+  const options = {
+    url: 'http://localhost:4090/Signin',
+    method: 'POST',
+    data: payload
+  };
+  return function (dispatch) {
+    axios(options)
+      .then(response => {
+        if (response.data === "User does not exist") {
+          alert("User does not exist")
+        }
+        else if (response.data === "wrong password") {
+          alert("Wrong password")
+        }
+        else {
+          sessionStorage.setItem('authentication', response.data.token)
+          BrowserHistory.push('/dashboard')
+        }
+      })
+      .catch(error => {
+        window.alert(error.response.data.message)
+      })
 
-return function(dispatch)
-{
-    console.log(payload)
-axios(options)
-.then(response => {
-    // if (response.data.msg === 'success') {
-    //     alert("Message sent successfully.");
-    // }
-console.log(response);
-sessionStorage.setItem('authentication', response.data.token)
-BrowserHistory.push('/dashboard')
-});
-dispatch({type:'Login',payload:payload});
+  }
 }
-}
+//     console.log(payload)
+// axios(options)
+// .then(response => {
 
+// console.log(response);
+// sessionStorage.setItem('authentication', response.data.token)
+// BrowserHistory.push('/dashboard')
+// });
+// dispatch({type:'Login',payload:payload});
+// }
+// }
